server/requirements.txt
django==3.2.14
djangorestframework==3.12.4
django-filter==21.1
django-mysql==4.2.0
mysqlclient==2.0.3
pytest==6.2.5
pytest-django==4.4.0
redis==4.0.2
requests>=2.27.0
gunicorn==20.1.0
gevent==21.12.0
Pillow==9.2.0
supervisor==4.2.4
rest-swagger==1.0.0
web3==5.28.0
PyMuPDF==1.19.6
requests-toolbelt==0.9.1
django-guardian==2.4.0
celery==5.2.6
django-apscheduler==0.6.2
django-grappelli==3.0.3
django-cors-headers==3.11.0
pytest-cov==3.0.0
tweepy==4.10.0
weasyprint==52.5